moved configuration and properties to dedicate package
